Run foreground service during Device Suspend Tests
The Device Suspend Tests verify that wake-up sensor events trigger a
device wake-up and the events are delivered to clients. Since these
tests require that the device enter the suspend state, it is possible
that the test application becomes idle while waiting for sensor
events. In Android P, applications that are idle are not able to
receive sensor events. Thus, if the test application becomes idle, it
will not receive the expected wake-up events, causing the tests to
erroneously fail.

This patch ensures that the test application maintains a foreground
service while it is running to ensure the application does not enter
the idle state and is able to receive sensor events.

Bug: 122817140
Test: Ensured the test application never enters the background state
      while the Device Suspend Tests are active.

Change-Id: I25944cc64a7471f2b5a6f3f070a6cbb6861c64fb

Former-commit-id: 9b724ed31f0e2a9586bf58c857fbaae9f8501ce7